requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
dictdiffer==0.8.0
google-auth==1.6.3
httplib2==0.13.0
idna==2.8
ipaddress==1.0.22
Jinja2==2.10.1
kubernetes==9.0.0
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.0.1
openshift==0.9.0rc0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
python-dateutil==2.8.0
python-string-utils==0.6.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
ruamel.ordereddict==0.4.13
ruamel.yaml==0.15.97
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
